==> notebooks/04_generating_historical_data.ipynb <==
# Importing the packages

import numpy as np
import pandas as pd
import seaborn as sns
import dask.dataframe as dd
import matplotlib.pyplot as plt

from sklearn import set_config
from sklearn.cluster import MiniBatchKMeans
from sklearn.preprocessing import StandardScaler
from sklearn.metrics import mean_absolute_percentage_error

sns.set_style("whitegrid")
set_config(transform_output='pandas')

# Loading the data

df_jan_path = "../data/raw/yellow_tripdata_2016-01.csv"
df_feb_path = "../data/raw/yellow_tripdata_2016-02.csv"
df_mar_path = "../data/raw/yellow_tripdata_2016-03.csv"

usecols = [
    "trip_distance", "tpep_pickup_datetime", "pickup_longitude", "pickup_latitude", "dropoff_longitude", "dropoff_latitude", "fare_amount"
]

df_jan = dd.read_csv(df_jan_path, assume_missing=True, usecols=usecols, parse_dates=["tpep_pickup_datetime"])
df_feb = dd.read_csv(df_feb_path, assume_missing=True, usecols=usecols, parse_dates=["tpep_pickup_datetime"])
df_mar = dd.read_csv(df_mar_path, assume_missing=True, usecols=usecols, parse_dates=["tpep_pickup_datetime"])

df_jan

## Concatenating the data

df = dd.concat([df_jan, df_feb, df_mar], axis=0)
df

# Bounds

min_lat = 40.60
max_lat = 40.85
min_lon = -74.05
max_lon = -73.70

min_fare_amount = 0.50
max_fare_amount = 81.0

min_trip_distance = 0.25
max_trip_distance = 24.43

# Removing outliers and unnecessary columns

df = df.loc[
    (df["pickup_latitude"].between(min_lat, max_lat, inclusive="both")) & 
    (df["pickup_longitude"].between(min_lon, max_lon, inclusive="both")) & 
    (df["dropoff_latitude"].between(min_lat, max_lat, inclusive="both")) & 
    (df["dropoff_longitude"].between(min_lon, max_lon, inclusive="both")),
    :
]
df = df.loc[
    (df["fare_amount"].between(min_fare_amount, max_fare_amount, inclusive="both")) & 
    (df["trip_distance"].between(min_trip_distance, max_trip_distance, inclusive="both")),
    :
]

df = df.drop(columns=["trip_distance", "dropoff_longitude", "dropoff_latitude", "fare_amount"])
df

# Saving the dataframe

df = df.compute()

INTERIM_DATA_PATH = "../data/interim/processing_data.csv"
# df.to_csv(INTERIM_DATA_PATH, index=False)

# Reading the saved dataframe and clustering

df_reader = pd.read_csv(INTERIM_DATA_PATH, chunksize=100000, usecols=["pickup_latitude", "pickup_longitude"])

scaler = StandardScaler()

for chunk in df_reader:
    scaler.partial_fit(chunk)

scaler

df_reader = pd.read_csv(INTERIM_DATA_PATH, chunksize=100000, usecols=["pickup_latitude", "pickup_longitude"])

mini_batch_k_means = MiniBatchKMeans(n_clusters=30, n_init=20, random_state=42)

for chunk in df_reader:
    scaled_chunk = scaler.transform(chunk)
    mini_batch_k_means.partial_fit(scaled_chunk)

mini_batch_k_means

scaler.inverse_transform(mini_batch_k_means.cluster_centers_)

location_subset = df[df.columns[1:]]
location_subset

scaled_location_subset = scaler.transform(location_subset)
scaled_location_subset

cluster_predictions = mini_batch_k_means.predict(scaled_location_subset)
cluster_predictions.shape

# save the cluster predictions in data

df['region'] = cluster_predictions
df

This has assigned a region to each row in the data according to the clusters.

# Dropping the latitude and longitude columns

time_series_data = df.drop(columns=["pickup_latitude", "pickup_longitude"])
time_series_data

time_series_data.dtypes

# Saving the time series data

TIME_SERIES_DATA_PATH = "../data/interim/time_series.csv"

# time_series_data.to_csv(TIME_SERIES_DATA_PATH, index=False)

# Setting the time series column as the index

time_series_data.set_index("tpep_pickup_datetime", inplace=True)
time_series_data

# Checking for missing values

time_series_data.isna().sum()

# Grouping the data by regions

region_grp = time_series_data.groupby("region")
region_grp

# Resampling the time series in 15 min intervals

resampled_time_series_data = (
    region_grp["region"]
    .resample("15min")
    .count()
)
resampled_time_series_data.name = "total_pickups"
resampled_time_series_data = resampled_time_series_data.reset_index(level=0)
resampled_time_series_data

# Checking number of zeros in the data

(resampled_time_series_data["total_pickups"] == 0).sum()

There are 3,668 15-minute time intervals that have 0 pickups. I will replace them with an arbitrary value. I did this because the `mean_absolute_percentage_error` function from sklearn does not perform well with 0 or close to 0 values in `y_true`.

epsilon_val = 10
resampled_time_series_data.replace({'total_pickups': {0 : epsilon_val}}, inplace=True)

(resampled_time_series_data["total_pickups"] == 0).sum()

# Smoothing

## Moving average

window_sizes = list(range(3, 11, 1))
window_sizes

def calculate_best_window_size(window_sizes: list):
    for window in window_sizes:
        idx = window - 1
        y_pred = resampled_time_series_data["total_pickups"].rolling(window=window).mean().values[idx:]
        y_true = resampled_time_series_data["total_pickups"].values[idx:]
        mape = mean_absolute_percentage_error(y_true=y_true, y_pred=y_pred)
        print(f"For the window size {window}, MAPE is {mape*100:.2f}%.")

calculate_best_window_size(window_sizes)

One thing that is clear is that to predict number of pickups, I need to give more importance to the more recent values as compared to the values that are more in the past.

## EWMA

resampled_time_series_data["total_pickups"].ewm(alpha=0.9).mean()

alpha_values = np.arange(0.2, 1, 0.1)
alpha_values

def calculate_best_alpha_value(alpha_values):
    y_true = resampled_time_series_data["total_pickups"].values
    for alpha in alpha_values:
        y_pred = resampled_time_series_data["total_pickups"].ewm(alpha=alpha).mean()
        mape = mean_absolute_percentage_error(y_true=y_true, y_pred=y_pred)
        print(f"For alpha = {alpha:.1f}, MAPE is {mape*100:.2f}%.")

calculate_best_alpha_value(alpha_values)

Clearly, more the `alpha` value, better the MAPE. This again reinforces that I should give more weight to the recent values. I will choose $\alpha = 0.4$ because this value is supposed to be chosen taking into consideration the number of lag features in the data. I am going to choose 4 lag features. For $n$ lag features, the value of $\alpha$ is given by
$$
\alpha = \frac{2}{n + 1}
$$
Substituting $n = 4$, we get $\alpha = 0.4$.

# Applying smoothing

resampled_time_series_data["avg_pickups"] = resampled_time_series_data["total_pickups"].ewm(alpha=0.4).mean().round()
resampled_time_series_data

resampled_time_series_data.shape

# Saving the final data

FINAL_TIME_SERIES_DATA_PATH = "../data/interim/final_time_series_data.csv"
# resampled_time_series_data.to_csv(FINAL_TIME_SERIES_DATA_PATH, index=True)
