==> notebooks/05_training_baseline_model.ipynb <==
# Importing the packages

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

from sklearn import set_config
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.linear_model import LinearRegression
from sklearn.metrics import mean_absolute_percentage_error

set_config(transform_output="pandas")

# Loading the data

FINAL_DATA_PATH = "../data/interim/final_time_series_data.csv"

df = pd.read_csv(FINAL_DATA_PATH, parse_dates=["tpep_pickup_datetime"])
df

df.shape

df.dtypes

# Checking for missing values

df.isna().sum()

# Extracting features

df["day_of_week"] = df["tpep_pickup_datetime"].dt.day_of_week
df["month"] = df["tpep_pickup_datetime"].dt.month
# As we have data for only 3 months, this month feature may not be very useful for prediction.
# However, I will use this feature for train-test split

# Setting the datetime column as the index

df.set_index("tpep_pickup_datetime", inplace=True)
df

# Creating lag features

region_grp = df.groupby("region")
region_grp

# Shifting periods

periods = list(range(1, 5))
periods

# Generating lag features

lag_features = region_grp["total_pickups"].shift(periods)
lag_features

# Merging the lag features with the original df

data = pd.concat([lag_features, df], axis=1)
data

print(f"Shape of the data before concatenation: {df.shape}.")
print(f"Shape of the data after concatenation: {data.shape}.")

# Number of rows having missing values

data.isna().any(axis=1).sum()

# Dropping the rows with missing values

data.dropna(inplace=True)
data.isna().any(axis=1).sum()

mapper = {
    name: f"lag_{idx + 1}" for idx, name in enumerate(data.columns[0:4])
}
mapper

# Replace column names

data = data.rename(columns=mapper)
data

# Number of rows in each month

data["month"].value_counts()

data["month"].value_counts(normalize=True)

# Train-test split

train_df = data.loc[data["month"].isin([1, 2]), "lag_1": "day_of_week"]
test_df = data.loc[data["month"].isin([3]), "lag_1": "day_of_week"]

train_df

test_df

# Saving the training and test sets

TRAIN_DATA_PATH = "../data/processed/train.csv"
TEST_DATA_PATH = "../data/processed/test.csv"

# train_df.to_csv(TRAIN_DATA_PATH, index=True)
# test_df.to_csv(TEST_DATA_PATH, index=True)

# Input-output split

X_train = train_df.drop(columns=["total_pickups"])
y_train = train_df["total_pickups"]

X_train

y_train

X_test = test_df.drop(columns=["total_pickups"])
y_test = test_df["total_pickups"]

X_test

y_test

# Regression pipeline

encoder = ColumnTransformer(
    [
        ("ohe", OneHotEncoder(drop="first", sparse_output=False), ["region", "day_of_week"]),
    ],
    remainder="passthrough",
    n_jobs=-1,
    force_int_remainder_cols=False,
)
encoder

# Encoding training and test sets

X_train_encoded = encoder.fit_transform(X_train)
X_test_encoded = encoder.transform(X_test)

# Linear regression

regressor = LinearRegression()

regressor.fit(X_train_encoded, y_train)

# Predictions

y_pred_train = regressor.predict(X_train_encoded)
y_pred_test = regressor.predict(X_test_encoded)

# Error metrics

train_mape = mean_absolute_percentage_error(y_true=y_train, y_pred=y_pred_train)
test_mape = mean_absolute_percentage_error(y_true=y_test, y_pred=y_pred_test)

print(f"Train MAPE: {train_mape * 100:.2f}%.")
print(f"Test MAPE: {test_mape * 100:.2f}%.")
