==> notebooks/06_model_selection_hyp_tuning.ipynb <==
# Importing the packages

import tqdm
import optuna
import mlflow
import dagshub
import numpy as np
import pandas as pd

from xgboost import XGBRegressor
from sklearn import set_config
from sklearn.svm import SVR
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.linear_model import LinearRegression, Ridge
from sklearn.metrics import mean_absolute_percentage_error
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor

set_config(transform_output="pandas")

# MLFlow and DagsHub

mlflow.set_tracking_uri("https://dagshub.com/SushrutGaikwad/taxi-demand-prediction.mlflow")
dagshub.init(repo_owner='SushrutGaikwad', repo_name='taxi-demand-prediction', mlflow=True)

# Loading the data

TRAIN_DATA_PATH = "../data/processed/train.csv"
TEST_DATA_PATH = "../data/processed/test.csv"

train_df = pd.read_csv(TRAIN_DATA_PATH, parse_dates=["tpep_pickup_datetime"]).set_index("tpep_pickup_datetime")
test_df = pd.read_csv(TEST_DATA_PATH, parse_dates=["tpep_pickup_datetime"]).set_index("tpep_pickup_datetime")

train_df

test_df

## Checking for missing values

train_df.isna().sum()

test_df.isna().sum()

# Input-output split

X_train = train_df.drop(columns=["total_pickups"])
y_train = train_df["total_pickups"]

X_test = test_df.drop(columns=["total_pickups"])
y_test = test_df["total_pickups"]

X_train.head()

y_train.head()

X_test.head()

y_test.head()

# Encoding

encoder = ColumnTransformer(
    [
        ("ohe", OneHotEncoder(drop="first", sparse_output=False), ["region", "day_of_week"]),
    ],
    remainder="passthrough",
    n_jobs=-1,
    force_int_remainder_cols=False,
)
encoder

X_train_encoded = encoder.fit_transform(X_train)
X_test_encoded = encoder.transform(X_test)

# Model Selection & Hyperparameter Tuning

# Experiment name

mlflow.set_experiment("model_selection")

def objective(trial):
    # Start with the child run
    with mlflow.start_run(nested=True) as child:
        # Models and their hyperparameter search space
        list_of_models = ["LR", "RF", "GBR", "XGBR"]
        model_name = trial.suggest_categorical("model_name", list_of_models)

        # Going over the models and their hyperparameters
        if model_name == "LR":
            model = LinearRegression()

        elif model_name == "RF":
            n_estimators_rf = trial.suggest_int("n_estimators_rf", 10, 100, step=10)
            max_depth_rf = trial.suggest_int("max_depth_rf", 3, 10)
            model = RandomForestRegressor(
                n_estimators=n_estimators_rf,
                max_depth=max_depth_rf,
                random_state=42,
                n_jobs=-1
            )

        elif model_name == "GBR":
            n_estimators_gb = trial.suggest_int("n_estimators_gb", 10, 100, step=10)
            learning_rate_gb = trial.suggest_float("learning_rate_gb", 1e-4, 1e-1, log=True)
            model = GradientBoostingRegressor(
                n_estimators=n_estimators_gb,
                learning_rate=learning_rate_gb,
                random_state=42
            )

        elif model_name == "XGBR":
            n_estimators_xgb = trial.suggest_int("n_estimators_xgb", 10, 100, step=10)
            learning_rate_xgb = trial.suggest_float("learning_rate_xgb", 1e-4, 1e-1, log=True)
            max_depth_xgb = trial.suggest_int("max_depth_xgb", 3, 10)
            model = XGBRegressor(
                n_estimators=n_estimators_xgb,
                learning_rate=learning_rate_xgb,
                max_depth=max_depth_xgb
            )

        # Logging the model name
        mlflow.log_param("model_name", model_name)

        # Logging the model parameters
        mlflow.log_params(model.get_params())

        # Training the model
        model.fit(X_train_encoded, y_train)

        # Getting the predictions
        y_pred = model.predict(X_test_encoded)

        # Calculating MAPE
        mape = mean_absolute_percentage_error(y_true=y_test, y_pred=y_pred)

        # Logging MAPE
        mlflow.log_metric("mape", mape)

        return mape

# Optimizing the objective function to get the best model

with mlflow.start_run(run_name="best_model", nested=True) as parent:
    # Creating a study object
    study = optuna.create_study(study_name="model_selection", direction="minimize")
    # Optimizing the objective function
    study.optimize(func=objective, n_trials=50, n_jobs=-1)

    # Log the best parameters
    mlflow.log_params(study.best_params)
    # Log the best MAPE
    mlflow.log_metric("best_mape", study.best_value)

# Best MAPE

print(f"The best MAPE is {study.best_value * 100:.2f}%.")

# Best model

study.best_params

# Model counts during the optimization

study.trials_dataframe()["params_model_name"].value_counts()

# Training the best model

regressor = LinearRegression()
regressor.fit(X_train_encoded, y_train)

# Predictions

y_pred_train = regressor.predict(X_train_encoded) 
y_pred_test = regressor.predict(X_test_encoded)

# MAPE

mape_train = mean_absolute_percentage_error(y_true=y_train, y_pred=y_pred_train)
mape_test = mean_absolute_percentage_error(y_true=y_test, y_pred=y_pred_test)

print(f"The training MAPE is {mape_train*100:.2f}%.")
print(f"The test MAPE is {mape_test*100:.2f}%.")

regressor.coef_

Let us try ridge regression, which is a regularized version of linear regression.

def tune_ridge(trial):
    # Hyperparameter space
    alpha = trial.suggest_float("alpha", 30, 100)
    
    # Model
    ridge = Ridge(alpha=alpha, random_state=42)
    
    # Training
    ridge.fit(X_train_encoded, y_train)
    
    # Predicting
    y_pred = ridge.predict(X_test_encoded)
    
    # MAPE
    mape = mean_absolute_percentage_error(y_true=y_test, y_pred=y_pred)

    return mape

# Create study

study = optuna.create_study(study_name="tune_ridge", direction="minimize")

# optimize

study.optimize(func=tune_ridge, n_trials=100, n_jobs=-1, show_progress_bar=True)

# Best parameters

study.best_params

# Best MAPE

print(f"The best MAPE using ridge regression is {study.best_value*100:.2f}%.")

So we will go with the linear regression model. We won't do any hyperparameter tuning for it as it does not have any hyperparameters.
